renamed load_blob to blob_loader
